AAdding validations and improving code

+++ src/main/java/com/sap/controllers/CalendarController.java
@@ -3,6 +3,7 @@ package com.sap.controllers;
 import com.sap.models.*;
 import com.sap.service.DayService;
 import com.sap.service.PeriodService;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -44,6 +45,7 @@ public class CalendarController extends BaseController{
      */
     @RequestMapping(value = "/calendar/manage/{id}", method = RequestMethod.GET)
     public String PeriodPage(@PathVariable("id") int id, Model model){
+        if(this.notAuthorized(id, "period")){return "errors/403";}
         if(this.getPrincipalUser().getRoles().get(1).getRole().equals("ROLE_MEMBER")){
             model.addAttribute("member", true);
         }
@@ -61,6 +63,9 @@ public class CalendarController extends BaseController{
     public String DayPage(@PathVariable("id") int id, Model model){
         User principal = this.getPrincipalUser();
         model.addAttribute("day", dayService.getDayById(id));
+
+        if(this.notAuthorized(id, "day")){return "errors/403";}
+
         for(Role role: principal.getRoles()){
             if(new String(role.getRole()).equals("ROLE_OWNER")){
                 model.addAttribute("userDays", userDayService.listUserDays(id));
@@ -124,6 +129,7 @@ public class CalendarController extends BaseController{
     @RequestMapping(value = "/day/admin/update", method = RequestMethod.POST)
     public String updateDay(WebRequest request, Model model){
         int id = Integer.parseInt(request.getParameter("id"));
+        if(this.notAuthorized(id, "day")){return "errors/403";}
 
         try{
             Day day = dayService.getDayById(id);
@@ -156,16 +162,12 @@ public class CalendarController extends BaseController{
         LocalDate start = LocalDate.parse(request.getParameter("start-date"));
         LocalDate end = LocalDate.parse(request.getParameter("end-date"));
 
-        if(start.isBefore(LocalDate.now()) || start.isAfter(end)){
+        if(!validatePeriods(periods, start, end)){
             model.addAttribute("error", "Not possible to create a period between this dates, check the corresponding values!");
             model.addAttribute("periods", periods);
             return "calendar-admin";
         }
-        if(!validatePeriods(periods, start, end)){
-            model.addAttribute("error", "Not possible to create the period, some days between the chosen interval has already been used!");
-            model.addAttribute("periods", periods);
-            return "calendar-admin";
-        }
+
         try{
             Period period = new Period();
             period.setStart(start);
@@ -220,6 +222,7 @@ public class CalendarController extends BaseController{
      * @return Boolean which symbolizes success or error
      */
     public Boolean validatePeriods(List<Period> periods, LocalDate start, LocalDate end){
+        if(start.isBefore(LocalDate.now()) || start.isAfter(end)){return false;}
         for(Period period : periods){
             if (start.isEqual(period.getStart())
                     || start.isEqual(period.getEnd())
@@ -242,10 +245,32 @@ public class CalendarController extends BaseController{
      */
     @RequestMapping("/calendar/delete/{id}")
     public String removePeriod(@PathVariable("id") int id, Model model){
+        if(this.notAuthorized(id, "period")){return "errors/403";}
+
         User principal = this.getPrincipalUser();
         periodService.removePeriod(id);
         model.addAttribute("periods", periodService.listPeriods(principal.getTeam().getId()));
         model.addAttribute("msg", "Period removed successfully!");
         return "calendar-admin";
+    }
+
+    public Boolean notAuthorized(int id, String type) {
+        User principal = this.getPrincipalUser();
+
+        if(type.equals("day")){
+            Day day = this.dayService.getDayById(id);
+            if (principal.getTeam().getId() != day.getPeriod().getTeam().getId()) {
+                return true;
+            }
+        }
+
+        if(type.equals("period")){
+            Period period = this.periodService.getPeriodById(id);
+            if (principal.getTeam().getId() != period.getTeam().getId()) {
+                return true;
+            }
+        }
+
+        return false;
     }
 }
